# [Perf] Track thread-CPU-time for each request on server-side.