## Objective

The objective of this change is to demonstrate some realistic usages of L3-logging, which is currently done on the server-side of a simple client/server msg-send-recv application.

Rather than logging a plain 'Server msg' info, with some client details, we now also track the real-time / thread-CPU-time consumed on the server-side, for each client-op request. We only measure the block of code after msgrecv(), so don't track other scaffolding overheads.

The L3-logging call is now changed to also log this metric, which is a more natural use-case for this kind of logging API.

@gregthelaw - You can start reviewing this multi-part PR. I have reworked the implementation based on what we discussed yesterday and to prepare this change-set to be more inline with the `l3-perf-bm` proposal aired out in this discussion #43 . 

------

## Implementation

This PR has 3 commits in this order:

1. l3_dump.py: Add errors='ignore' clause to parse_rodata_string_offsets(): Fix applied to dump Python script to deal with some case where Unicode chars are generated in one of the test workloads.

2. [Perf] Track time-consumed for processing each request on server-side.

3. [Perf] Improve usability and metrics of client-server perf testbed: Tracks throughput metrics.

With the last 2 changes, following is the summary output w/o L3-logging (baseline) and with L3-logging:

```
agurajada-Linux-Vm:[3152] $ egrep -E 'Start Server|clients, num_ops' ~/tmp/perf.1Mil.out

Start Server, using clock ID=3 'CLOCK_THREAD_CPUTIME_ID': No logging.
For 5 clients, num_ops=5000000 (5 Million) ops, Avg. Thread-CPU time=542 ns/msg, throughput=1843753 (~1.84 Million) ops/sec

Start Server, using clock '/tmp/l3.c-server-test.dat': Initiate L3-logging to log-file '(no LOC)', using CLOCK_THREAD_CPUTIME_ID encoding scheme.
For 5 clients, num_ops=5000000 (5 Million) ops, Avg. Thread-CPU time=587 ns/msg, throughput=1703080 (~1.70 Million) ops/sec
```

-------

This commit implements on the server-side:

- Thread-specific CPU-time consumed for processing each client request received.
- Cumulative CPU-time and # requests is also tracked for each client.
- L3-log is changed to log the CPU-time consumed to implement each request.
- When client exits, the stats of # of requests received and avg. CPU-time/request is printed.

Measurement clock chosen:

 - Linux: Use CLOCK_REALTIME as it seems to have a resolution of 1 ns.
 - Mac: We use CLOCK_THREAD_CPUTIME_ID to measure the time taken to implement
   each request as its resolution is 1ns. All other clocks have a resolution
   of 1000 ns (as seen on my Mac).

With this enhancement, we should see these enhanced messages upon running the
L3-dump after running the exercise:

- Refactor `common timespec_to_ns()` into `svmsg_file.h` as it's now shared by client / server files.

We should see these enhanced messages (on a Mac/OSX):

- On server-side:
```
Server: Client ID=5701633 exited. num_ops=1000, Avg. thread-CPU-ns=351 ns/msg, # active clients=2
Server: Client ID=5570563 exited. num_ops=1000, Avg. thread-CPU-ns=351 ns/msg, # active clients=1
Client: ID=5570564 Performed 1000 message send/receive operations, ctr=1000, 35803 ns/msg (avg) Exiting.
Server: Client ID=5570564 exited. num_ops=1000, Avg. thread-CPU-ns=352 ns/msg, # active clients=0
Server: # active clients=0 (HWM=5). Exiting.
```

- From L3-dump utility:
```
tid=259 'Server msg: Increment: ClientID=5570563, Thread CPU-time=326 ns.' tid=259 'Server msg: Increment: ClientID=5570564, Thread CPU-time=361 ns.' tid=259 'Server msg: Increment: ClientID=5570564, Thread CPU-time=446 ns.' tid=259 'Server msg: Increment: ClientID=5570564, Thread CPU-time=453 ns.' Unpacked nentries=5000 log-entries.
```


---
With L3_LOC enabled L3-log'ged messages are like follows:

---
tid=259 client-server-msgs-perf/svmsg_file_server.c:258  'Server msg: Increment: ClientID=5636100, Thread CPU-time=333 ns.'
tid=259 client-server-msgs-perf/svmsg_file_server.c:258  'Server msg: Increment: ClientID=5636098, Thread CPU-time=331 ns.'
tid=259 client-server-msgs-perf/svmsg_file_server.c:258  'Server msg: Increment: ClientID=5767169, Thread CPU-time=334 ns.'
tid=259 client-server-msgs-perf/svmsg_file_server.c:258  'Server msg: Increment: ClientID=5636100, Thread CPU-time=329 ns.'
tid=259 client-server-msgs-perf/svmsg_file_server.c:258  'Server msg: Increment: ClientID=5767169, Thread CPU-time=446 ns.'
tid=259 client-server-msgs-perf/svmsg_file_server.c:258  'Server msg: Increment: ClientID=5636100, Thread CPU-time=340 ns.'
tid=259 client-server-msgs-perf/svmsg_file_server.c:258  'Server msg: Increment: ClientID=5767169, Thread CPU-time=425 ns.'
Unpacked nentries=5000 log-entries.

--
Note the `Thread CPU-time=340 ns.` fragment, which is new with this change.