@gapisback 2024-05-13
@gregthelaw - thanks for looking at this work.

Re: "_FWIW I'm quite disappointed that L3 adds so much overhead_" ... in [this comment](https://github.com/undoio/l3/pull/42#pullrequestreview-2051429233), here's my take.

I look at this change-set as a stable working implementation on which we can do more u-benchmarking experiments.

Do note that I have only run L3-logging in "slow" mode, i.e., it's a function call to a C-function, not the assembly code. I'll see if I can somehow stitch a perf-version where one can select slow / fast logging api.

I am running this on a Linux-VM running on my Mac, so there probably are other overheads.

Also, the perf-experimentations for `fprintf()`, spdlog will happen soon, but for that I'd like this core infra piece checked-in.

I think once we get this work stable, it will be good to port this to real h/w or a VM running on some cloud which may have more stable performance behaviour. (Let's take this offline.)

I'll give this a final look and will prepare for check-in to `/main`.